fix(dataset): ensuring firewall ok before every operation
