+++ src/Sepes.Functions/Startup.cs
@@ -121,6 +121,8 @@ namespace Sepes.CloudResourceWorker
             builder.Services.AddTransient<IAzureVmService, AzureVmService>();
             builder.Services.AddTransient<IAzureQueueService, AzureQueueService>();
             builder.Services.AddTransient<IAzureStorageAccountService, AzureStorageAccountService>();
+            builder.Services.AddTransient<IAzureStorageAccountCorsRuleService, AzureStorageAccountCorsRuleService>();
+            builder.Services.AddTransient<IAzureStorageAccountNetworkRuleService, AzureStorageAccountNetworkRuleService>();
             builder.Services.AddTransient<IAzureCostManagementService, AzureCostManagementService>();
             builder.Services.AddTransient<IAzureResourceSkuService, AzureResourceSkuService>();
             builder.Services.AddTransient<IAzureDiskPriceService, AzureDiskPriceService>();
