+++ src/Sepes.Infrastructure/Service/Azure/Interface/IAzureStorageAccountCorsRuleService.cs
@@ -1,0 +1,7 @@
+﻿namespace Sepes.Infrastructure.Service.Azure.Interface
+{
+    public interface IAzureStorageAccountCorsRuleService : IHasCorsRules
+    {
+      
+    }
+}
