+++ src/Sepes.Infrastructure/Service/Azure/Interface/IAzureStorageAccountNetworkRuleService.cs
@@ -1,0 +1,11 @@
+﻿using System.Threading;
+using System.Threading.Tasks;
+
+namespace Sepes.Infrastructure.Service.Azure.Interface
+{
+    public interface IAzureStorageAccountNetworkRuleService : IHasFirewallRules
+    {
+        Task AddStorageAccountToVNet(string resourceGroupForStorageAccount, string storageAccountName, string resourceGroupForVnet, string vNetName, CancellationToken cancellation);
+        Task RemoveStorageAccountFromVNet(string resourceGroupForStorageAccount, string storageAccountName, string resourceGroupForVnet, string vNetName, CancellationToken cancellation);
+    }
+}
