+++ src/Sepes.Infrastructure/Service/Azure/Interface/IAzureStorageAccountService.cs
@@ -6,14 +6,13 @@ using System.Threading.Tasks;
 
 namespace Sepes.Infrastructure.Service.Azure.Interface
 {
-    public interface IAzureStorageAccountService : IHasProvisioningState, IHasTags, IPerformResourceProvisioning, IHasNetworkRules, IHasCorsRules
+    public interface IAzureStorageAccountService : IHasProvisioningState, IHasTags, IPerformResourceProvisioning
     {
         Task Delete(string resourceGroupName, string storageAccountName, CancellationToken cancellationToken = default);
 
         Task<AzureStorageAccountDto> Create(Region region, string resourceGroupName, string storageAccountName, Dictionary<string, string> tags, List<string> onlyAllowAccessFrom = null, CancellationToken cancellationToken = default);
 
 
-        Task AddStorageAccountToVNet(string resourceGroupForStorageAccount, string storageAccountName, string resourceGroupForVnet, string vNetName, CancellationToken cancellation);
-        Task RemoveStorageAccountFromVNet(string resourceGroupForStorageAccount, string storageAccountName, string resourceGroupForVnet, string vNetName, CancellationToken cancellation);
+      
     }
 }
