+++ src/Sepes.Infrastructure/Service/Azure/StorageAccount/AzureStorageAccountTokenService.cs
