+++ src/Sepes.Infrastructure/Util/Azure/AzureResourceServiceResolver.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using Sepes.Infrastructure.Constants;
 using Sepes.Infrastructure.Service;
+using Sepes.Infrastructure.Service.Azure;
 using Sepes.Infrastructure.Service.Azure.Interface;
 using System;
 
@@ -54,6 +55,19 @@ namespace Sepes.Infrastructure.Util
 
             return service;
         }
+
+        public static IHasCorsRules GetCorsRuleServiceOrThrow(IServiceProvider serviceProvider, string resourceType) => resourceType switch
+        {           
+            AzureResourceType.StorageAccount => serviceProvider.GetRequiredService<IAzureStorageAccountCorsRuleService>(),    
+            _ => throw new NullReferenceException($"Unable to cors provisioning service for type {resourceType}"),
+        };
+
+        public static IHasFirewallRules GetFirewallRuleService(IServiceProvider serviceProvider, string resourceType) => resourceType switch
+        {
+            AzureResourceType.StorageAccount => serviceProvider.GetRequiredService<IAzureStorageAccountNetworkRuleService>(),
+            _ => throw new NullReferenceException($"Unable to resolve firewall service for type {resourceType}"),
+        };
+
     }
 }
