+++ src/Sepes.Infrastructure/Util/Provisioning/EnsureFirewallRulesUtil.cs
@@ -27,7 +27,7 @@ namespace Sepes.Infrastructure.Util.Provisioning
 
         public static async Task Handle(
             CloudResourceOperationDto operation,
-            IHasNetworkRules networkRuleService,
+            IHasFirewallRules networkRuleService,
             ICloudResourceReadService resourceReadService,
             ICloudResourceOperationUpdateService operationUpdateService,
             ILogger logger)
@@ -43,7 +43,7 @@ namespace Sepes.Infrastructure.Util.Provisioning
 
                 var rulesFromOperationState = CloudResourceConfigStringSerializer.DesiredFirewallRules(operation.DesiredState);
 
-                var setRulesTask = networkRuleService.SetNetworkRules(operation.Resource.ResourceGroupName, operation.Resource.ResourceName, rulesFromOperationState, cancellation.Token);
+                var setRulesTask = networkRuleService.SetFirewallRules(operation.Resource.ResourceGroupName, operation.Resource.ResourceName, rulesFromOperationState, cancellation.Token);
 
                 while (!setRulesTask.IsCompleted)
                 {
